+++ types/Utils.d.ts
@@ -1,8 +1,4 @@
-export enum SearchTypes {
-    soundcloud = 'scsearch',
-    youtube = 'ytsearch',
-    youtubemusic = 'ytmsearch'
-}
+import { SearchTypes } from "./Constants";
 
 export class Utils {
     public static getVersion(): { variant: 'light' | 'vanilla', version: string } | never;
